Refactor: index.js를 render.js와 router.js로 분리
