+++ static/index.html
@@ -475,6 +475,6 @@
         </article>
       </main>
     </div>
-    <script type="module" src="/static/js/index.js"></script>
+    <script type="module" src="/static/js/init.js"></script>
   </body>
 </html>
